==> 1_Data.ipynb <==
## 1. Load Data

Load all reviews

import pandas as pd

reviews = pd.read_csv("~/Downloads/yelp-2019-dataset/review-005.csv", usecols=['business_id', 'text', 'stars'])
reviews

Load all businesses

businesses = pd.read_csv('~/Downloads/yelp-2019-dataset/business.csv', usecols=['business_id', 'categories'])
businesses

## 2. Filter / Clean Data

Filter businesses with "Restaurants" category

businesses = businesses[businesses['categories'].notnull()]
businesses["categories"] = businesses["categories"].str.split(", ")
businesses = businesses[businesses['categories'].map(lambda categories: 'Restaurants' in categories)]

businesses

Filter reviews to only those that were given to Restaurant businesses

reviews = reviews.merge(businesses, on='business_id', how='inner')

Shuffle the filtered reviews

reviews = reviews.sample(frac=1, random_state=42).reset_index(drop=True)

Drop reviews without text

reviews = reviews.dropna()

Reviews after filtered

reviews

Statistics of review star ratings

reviews.describe().transpose()

Distribution of review star ratings in proportion

reviews.stars.value_counts(normalize=True).sort_index()

Distribution of review star ratings in counts

reviews.stars.value_counts().sort_index()

## Data sampling

Samples 500,000 reviews

reviews = reviews[0:500000]

reviews

Statistics of review star ratings (sampled)

reviews.describe().transpose()

Distribution of review star ratings in proportion (sampled)

reviews.stars.value_counts(normalize=True).sort_index()

Distribution of review star ratings in counts (sampled)

reviews.stars.value_counts().sort_index()

## Save Data

reviews.to_csv('./data/reviews_500k_imba.csv', index=False)
