==> 2_Text-Preprocessing.ipynb <==
# Setup

Install required libraries

!pip install contractions

Import required libraries

import numpy as np
import pandas as pd

Mount storage from Google Drive

from google.colab import drive
drive.mount('p2')

# Dataset

Load sampled dataset

df = pd.read_csv('/content/p2/MyDrive/p2/data/reviews_500k_imba.csv')
df.head()

Inspect distribution of star labels

df['stars'].value_counts()

# Preprocessing

Remove unused columns

df = df[['text', 'stars']]
df

Minus 1 so the star labels range from 0-4 to prepare for one-hot encodings

df.loc[:,'stars'] -= 1
df

The example text before prepocessing

df['text'][1]

Expand contractions for the review text. Eg: I'm to I am

import contractions

df.loc[440678, 'text'] = df['text'][440678].replace("İ", "I") # The special symbol in this specific row is causing error when expanding
df['processed_text'] = [contractions.fix(text) for text in df['text']]
df['processed_text'][1]

Convert to lowercase and remove extra space

import nltk
nltk.download('punkt')

%pprint

df['processed_text'] = [nltk.word_tokenize(text.lower()) for text in df['processed_text']]

[w for w in df['processed_text'][1]]

Remove punctuations

df['processed_text'] = [[token for token in tokens if token.isalnum()] for tokens in df.processed_text]
' '.join(df['processed_text'][1])

Mark the immediate word after negation

negation_words = ['not', 'no', 'never', 'none', 'nobody', 'nothing', 'neither', 'nowhere']

for tokens in df.processed_text:
  for index, token in enumerate(tokens):
    if token in negation_words and index != len(tokens)-1:
      tokens[index + 1] = "!" + tokens[index + 1]

' '.join(df['processed_text'][1])

Remove English stop words

nltk.download('stopwords')

stopwords = nltk.corpus.stopwords.words("english")
df['processed_text'] = [[token for token in tokens if token not in stopwords] for tokens in df.processed_text]
' '.join(df['processed_text'][1])

Lemmatize tokens

nltk.download('wordnet')

from nltk.stem import WordNetLemmatizer

lemmatizer = WordNetLemmatizer()

df['processed_text'] = [[lemmatizer.lemmatize(token) for token in tokens] for tokens in df.processed_text]

' '.join(df['processed_text'][1])

Join the processed tokens back into text

df['processed_text'] = [str.join(' ', tokens) for tokens in df.processed_text]
df['processed_text'][1]

Save the preprocessed data frame for future use

df.to_csv('/content/p2/MyDrive/p2/data/preprocessed_500k_imba.csv', index=False)
