==> 4-1_Multinomial-Logistic-Regression.ipynb <==
## 1. Setup & Load dataset

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

from google.colab import drive
drive.mount('p2')

df = pd.read_csv('/content/p2/MyDrive/p2/data/preprocessed_500k_imba.csv')
df.fillna('', inplace=True)
df.head()

## 2. Preprocess

%pprint
df['processed_text'] = [text.split(' ') for text in df.processed_text]
[w for w in df['processed_text'][1]]

## 3. Train, Validation, Test Split

from sklearn.model_selection import train_test_split

x_train, x_test, y_train, y_test = train_test_split(
    df['processed_text'], 
    df['stars'], 
    test_size=0.2, 
    stratify=df['stars'], 
    random_state=42
)

x_train, x_val, y_train, y_val = train_test_split(
    x_train, 
    y_train, 
    test_size=0.25, 
    stratify=y_train, 
    random_state=42
)

print(x_train.shape)
print(x_val.shape)
print(x_test.shape)

## 4. Vectorization & Modeling

### 4.1 Define functions

from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.linear_model import LogisticRegression
from sklearn.metrics import classification_report

def identity_tokenizer(text):
    return text

def run_classifier(tfidf, model):
  X_train = tfidf.fit_transform(x_train)
  X_val = tfidf.transform(x_val)

  print("X_train shape:", X_train.shape)
  print("X_val shape:", X_val.shape)

  model.fit(X_train, y_train)

  y_pred = model.predict(X_val)
  print("Evaluation results:")
  print(classification_report(y_val, y_pred, digits=4))

### 4.2 Base Model

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=10000, ngram_range=(1,2))
model = LogisticRegression(multi_class='multinomial', solver='newton-cg')

run_classifier(tfidf, model)

### 4.3 Vectorizer Tunings

#### 4.3.1 Max Features = 5000

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=5000, ngram_range=(1,2))
model = LogisticRegression(multi_class='multinomial', solver='newton-cg')

run_classifier(tfidf, model)

#### 4.3.3 Max Features = 15000

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=15000, ngram_range=(1,2))
model = LogisticRegression(multi_class='multinomial', solver='newton-cg')

run_classifier(tfidf, model)

#### 4.3.3 Unigram

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=10000, ngram_range=(1,1))
model = LogisticRegression(multi_class='multinomial', solver='newton-cg')

run_classifier(tfidf, model)

#### 4.3.4 Bigram

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=10000, ngram_range=(2,2))
model = LogisticRegression(multi_class='multinomial', solver='newton-cg')

run_classifier(tfidf, model)

#### 4.3.5 Max Document Frequency = 0.75

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=10000, ngram_range=(1,2), max_df=0.75)
model = LogisticRegression(multi_class='multinomial', solver='newton-cg')

run_classifier(tfidf, model)

#### 4.3.6 Max Document Frequency = 0.50

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=10000, ngram_range=(1,2), max_df=0.5)
model = LogisticRegression(multi_class='multinomial', solver='newton-cg')

run_classifier(tfidf, model)

### 4.4 Model Hyperparameter Tunings

#### 4.4.1 C-value = 0.5

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=10000, ngram_range=(1,2))
model = LogisticRegression(multi_class='multinomial', solver='newton-cg', C=0.5)

run_classifier(tfidf, model)

#### 4.4.2 C-value = 0.1

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=10000, ngram_range=(1,2))
model = LogisticRegression(multi_class='multinomial', solver='newton-cg', C=0.1)

run_classifier(tfidf, model)

#### Solver = lbfgs

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=10000, ngram_range=(1,2))
model = LogisticRegression(multi_class='multinomial', solver='lbfgs')

run_classifier(tfidf, model)

#### Solver = saga

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=10000, ngram_range=(1,2))
model = LogisticRegression(multi_class='multinomial', solver='saga')

run_classifier(tfidf, model)

## 5. Evaluate

from sklearn.metrics import confusion_matrix

tfidf = TfidfVectorizer(tokenizer=identity_tokenizer, lowercase=False, max_features=15000, ngram_range=(1,2))
model = LogisticRegression(multi_class='multinomial', solver='newton-cg', C=0.5)

X_train = tfidf.fit_transform(x_train)
X_test = tfidf.transform(x_test)

print("X_train shape:", X_train.shape)
print("X_test shape:", X_test.shape)

model.fit(X_train, y_train)

y_pred = model.predict(X_test)
print("Evaluation results:")
print(classification_report(y_test, y_pred, digits=4))
print(confusion_matrix(y_test, y_pred))
